RouteUrl: move to internal package
